consumer/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = ["Bart van Arnhem <bart@orbisk.com>"]

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"
urllib3 = "^2.0.2"
prometheus_client = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
